+++ core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/HelperTool.java
@@ -6,7 +6,6 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.List;
-import java.util.function.Function;
 import java.util.stream.Stream;
 
 import org.eclipse.rdf4j.model.Model;
@@ -15,8 +14,6 @@ import org.eclipse.rdf4j.model.Statement;
 import org.eclipse.rdf4j.model.Value;
 import org.eclipse.rdf4j.model.vocabulary.RDF;
 import org.eclipse.rdf4j.repository.RepositoryConnection;
-import org.eclipse.rdf4j.sail.shacl.ast.constraintcomponents.ConstraintComponent;
-import org.eclipse.rdf4j.sail.shacl.ast.planNodes.ValidationTuple;
 
 public class HelperTool {
 
@@ -81,122 +78,6 @@ public class HelperTool {
 		}
 
 		return ret;
-
-	}
-
-	public static class SubjectScopedMapper implements Function<Statement, ValidationTuple> {
-
-		private final ConstraintComponent.Scope scope;
-
-		private SubjectScopedMapper(ConstraintComponent.Scope scope) {
-			this.scope = scope;
-		}
-
-		static SubjectScopedMapper nodeShapeInstance = new SubjectScopedMapper(ConstraintComponent.Scope.nodeShape);
-		static SubjectScopedMapper propertyShapeInstance = new SubjectScopedMapper(
-				ConstraintComponent.Scope.propertyShape);
-		static SubjectScopedMapper noneInstance = new SubjectScopedMapper(ConstraintComponent.Scope.none);
-
-		@Override
-		public ValidationTuple apply(Statement s) {
-			return new ValidationTuple(s.getSubject(), scope, false);
-		}
-
-		@Override
-		public int hashCode() {
-			return 72706357 + scope.hashCode();
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			return obj == this;
-		}
-
-		public static SubjectScopedMapper getFunction(ConstraintComponent.Scope scope) {
-			switch (scope) {
-
-			case none:
-				return noneInstance;
-			case nodeShape:
-				return nodeShapeInstance;
-			case propertyShape:
-				return propertyShapeInstance;
-			}
-
-			throw new IllegalStateException("Unknown scope: " + scope);
-		}
-
-	}
-
-	public static class ObjectScopedMapper implements Function<Statement, ValidationTuple> {
-
-		private final ConstraintComponent.Scope scope;
-
-		private ObjectScopedMapper(ConstraintComponent.Scope scope) {
-			this.scope = scope;
-		}
-
-		static ObjectScopedMapper nodeShapeInstance = new ObjectScopedMapper(ConstraintComponent.Scope.nodeShape);
-		static ObjectScopedMapper propertyShapeInstance = new ObjectScopedMapper(
-				ConstraintComponent.Scope.propertyShape);
-		static ObjectScopedMapper noneInstance = new ObjectScopedMapper(ConstraintComponent.Scope.none);
-
-		@Override
-		public ValidationTuple apply(Statement s) {
-			return new ValidationTuple(s.getObject(), scope, false);
-		}
-
-		@Override
-		public int hashCode() {
-			return 25482634 + scope.hashCode();
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			return obj == this;
-		}
-
-		public static ObjectScopedMapper getFunction(ConstraintComponent.Scope scope) {
-			switch (scope) {
-
-			case none:
-				return noneInstance;
-			case nodeShape:
-				return nodeShapeInstance;
-			case propertyShape:
-				return propertyShapeInstance;
-			}
-
-			throw new IllegalStateException("Unknown scope: " + scope);
-		}
-
-	}
-
-	public static class SubjectObjectPropertyShapeMapper implements Function<Statement, ValidationTuple> {
-
-		private SubjectObjectPropertyShapeMapper() {
-		}
-
-		static SubjectObjectPropertyShapeMapper instance = new SubjectObjectPropertyShapeMapper();
-
-		@Override
-		public ValidationTuple apply(Statement s) {
-			return new ValidationTuple(s.getSubject(), s.getObject(), ConstraintComponent.Scope.propertyShape, true);
-		}
-
-		@Override
-		public int hashCode() {
-			return 35972357;
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			return obj == this;
-		}
-
-		public static SubjectObjectPropertyShapeMapper getFunction() {
-			return instance;
-		}
 
 	}
