+++ core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/planNodes/UnorderedSelect.java
@@ -20,6 +20,7 @@ import org.eclipse.rdf4j.model.Value;
 import org.eclipse.rdf4j.sail.SailConnection;
 import org.eclipse.rdf4j.sail.SailException;
 import org.eclipse.rdf4j.sail.memory.MemoryStoreConnection;
+import org.eclipse.rdf4j.sail.shacl.ast.constraintcomponents.ConstraintComponent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -93,6 +94,8 @@ public class UnorderedSelect implements PlanNode {
 		stringBuilder.append(getId() + " [label=\"" + StringEscapeUtils.escapeJava(this.toString()) + "\"];")
 				.append("\n");
 
+		// added/removed connections are always newly minted per plan node, so we instead need to compare the underlying
+		// sail
 		if (connection instanceof MemoryStoreConnection) {
 			stringBuilder
 					.append(System.identityHashCode(((MemoryStoreConnection) connection).getSail()) + " -> " + getId())
@@ -142,6 +145,8 @@ public class UnorderedSelect implements PlanNode {
 		}
 		UnorderedSelect that = (UnorderedSelect) o;
 
+		// added/removed connections are always newly minted per plan node, so we instead need to compare the underlying
+		// sail
 		if (connection instanceof MemoryStoreConnection && that.connection instanceof MemoryStoreConnection) {
 			return ((MemoryStoreConnection) connection).getSail()
 					.equals(((MemoryStoreConnection) that.connection).getSail()) &&
@@ -161,11 +166,132 @@ public class UnorderedSelect implements PlanNode {
 
 	@Override
 	public int hashCode() {
+		// added/removed connections are always newly minted per plan node, so we instead need to compare the underlying
+		// sail
 		if (connection instanceof MemoryStoreConnection) {
 			return Objects.hash(((MemoryStoreConnection) connection).getSail(), subject, predicate, object, mapper,
 					UnorderedSelect.class);
 		}
 
 		return Objects.hash(connection, subject, predicate, object, mapper, UnorderedSelect.class);
+	}
+
+	public static class Mapper {
+		public static class SubjectScopedMapper implements Function<Statement, ValidationTuple> {
+
+			private final ConstraintComponent.Scope scope;
+
+			private SubjectScopedMapper(ConstraintComponent.Scope scope) {
+				this.scope = scope;
+			}
+
+			static SubjectScopedMapper nodeShapeInstance = new SubjectScopedMapper(ConstraintComponent.Scope.nodeShape);
+			static SubjectScopedMapper propertyShapeInstance = new SubjectScopedMapper(
+					ConstraintComponent.Scope.propertyShape);
+			static SubjectScopedMapper noneInstance = new SubjectScopedMapper(ConstraintComponent.Scope.none);
+
+			@Override
+			public ValidationTuple apply(Statement s) {
+				return new ValidationTuple(s.getSubject(), scope, false);
+			}
+
+			@Override
+			public int hashCode() {
+				return 72706357 + scope.hashCode();
+			}
+
+			@Override
+			public boolean equals(Object obj) {
+				return obj == this;
+			}
+
+			public static SubjectScopedMapper getFunction(ConstraintComponent.Scope scope) {
+				switch (scope) {
+
+				case none:
+					return noneInstance;
+				case nodeShape:
+					return nodeShapeInstance;
+				case propertyShape:
+					return propertyShapeInstance;
+				}
+
+				throw new IllegalStateException("Unknown scope: " + scope);
+			}
+
+		}
+
+		public static class ObjectScopedMapper implements Function<Statement, ValidationTuple> {
+
+			private final ConstraintComponent.Scope scope;
+
+			private ObjectScopedMapper(ConstraintComponent.Scope scope) {
+				this.scope = scope;
+			}
+
+			static ObjectScopedMapper nodeShapeInstance = new ObjectScopedMapper(ConstraintComponent.Scope.nodeShape);
+			static ObjectScopedMapper propertyShapeInstance = new ObjectScopedMapper(
+					ConstraintComponent.Scope.propertyShape);
+			static ObjectScopedMapper noneInstance = new ObjectScopedMapper(ConstraintComponent.Scope.none);
+
+			@Override
+			public ValidationTuple apply(Statement s) {
+				return new ValidationTuple(s.getObject(), scope, false);
+			}
+
+			@Override
+			public int hashCode() {
+				return 25482634 + scope.hashCode();
+			}
+
+			@Override
+			public boolean equals(Object obj) {
+				return obj == this;
+			}
+
+			public static ObjectScopedMapper getFunction(ConstraintComponent.Scope scope) {
+				switch (scope) {
+
+				case none:
+					return noneInstance;
+				case nodeShape:
+					return nodeShapeInstance;
+				case propertyShape:
+					return propertyShapeInstance;
+				}
+
+				throw new IllegalStateException("Unknown scope: " + scope);
+			}
+
+		}
+
+		public static class SubjectObjectPropertyShapeMapper implements Function<Statement, ValidationTuple> {
+
+			private SubjectObjectPropertyShapeMapper() {
+			}
+
+			static SubjectObjectPropertyShapeMapper instance = new SubjectObjectPropertyShapeMapper();
+
+			@Override
+			public ValidationTuple apply(Statement s) {
+				return new ValidationTuple(s.getSubject(), s.getObject(), ConstraintComponent.Scope.propertyShape,
+						true);
+			}
+
+			@Override
+			public int hashCode() {
+				return 35972357;
+			}
+
+			@Override
+			public boolean equals(Object obj) {
+				return obj == this;
+			}
+
+			public static SubjectObjectPropertyShapeMapper getFunction() {
+				return instance;
+			}
+
+		}
 	}
 }
