+++ core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/targets/DashAllObjects.java
@@ -13,7 +13,6 @@ import org.eclipse.rdf4j.model.vocabulary.SHACL;
 import org.eclipse.rdf4j.sail.SailConnection;
 import org.eclipse.rdf4j.sail.shacl.ConnectionsGroup;
 import org.eclipse.rdf4j.sail.shacl.RdfsSubClassOfReasoner;
-import org.eclipse.rdf4j.sail.shacl.ast.HelperTool;
 import org.eclipse.rdf4j.sail.shacl.ast.StatementMatcher;
 import org.eclipse.rdf4j.sail.shacl.ast.constraintcomponents.ConstraintComponent;
 import org.eclipse.rdf4j.sail.shacl.ast.planNodes.ExternalFilterIsObject;
@@ -52,7 +51,7 @@ public class DashAllObjects extends Target {
 			SailConnection connection) {
 
 		return new Unique(new UnorderedSelect(connection, null,
-				null, null, HelperTool.ObjectScopedMapper.getFunction(scope)));
+				null, null, UnorderedSelect.Mapper.ObjectScopedMapper.getFunction(scope)));
 
 	}
